+++ CHANGELOG.md
@@ -4,3 +4,7 @@
 
 ## 1.0.1
 - Fix: Added example to README.md
+
+## 1.0.2
+- Fix: A example project was added to the package.
+- Test: Added tests for the package.